Clean up + renaming tests
